Content/requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
Django==4.0.1
django-cors-headers==3.11.0
django-mysql==4.5.0
djangorestframework==3.13.1
mysqlclient==2.1.0
Faker==12.0.0
pika==1.2.0
pytz==2021.3
sqlparse==0.4.2
tzdata==2021.5
